+++ lib/src/easy_localization/file_loaders/io_file_loader.dart
@@ -1,0 +1,2 @@
+export 'io_file_loader_stub.dart'
+    if (dart.library.io) 'io_file_loader_io.dart';