+++ lib/src/extensions/locale_extensions.dart
@@ -38,7 +38,16 @@ extension LocaleExtension on Locale {
 }
 
 extension XLocaleExtension on XLocale {
-  Locale get locale => Locale(languageCode, countryCode);
+  Locale get locale {
+    if (scriptCode != null && scriptCode!.isNotEmpty) {
+      return Locale.fromSubtags(
+        languageCode: languageCode,
+        countryCode: countryCode,
+        scriptCode: scriptCode,
+      );
+    }
+    return Locale(languageCode, countryCode);
+  }
 
   String toStringWithSeparator({String separator = '-'}) {
     return locale.toStringWithSeparator(separator: separator);